% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lab_tests_joined}
\alias{lab_tests_joined}
\title{lab_tests_joined}
\format{Tibble with 9492 observations and 15 attributes}
\usage{
lab_tests_joined
}
\description{
lab_tests_joined
}
\keyword{datasets}
